# Optimize decoding utf8 buffers to strings.