--- a/lib/utf8/index.js
+++ b/lib/utf8/index.js
@@ -75,23 +75,18 @@ function ascii_decode_unrolled(buffer, start, end, initialStr, initialPos) {
     var str = initialStr || "";
     var i = initialPos || start;
 
-    // process 8 bytes at a time when possible
-    for (; i + 7 < end; i += 8) {
-        // Check all 8 bytes at once for non-ASCII using bitwise OR
-        if (buffer[i] > 0x7F || buffer[i + 1] > 0x7F || buffer[i + 2] > 0x7F || buffer[i + 3] > 0x7F || buffer[i + 4] > 0x7F || buffer[i + 5] > 0x7F || buffer[i + 6] > 0x7F || buffer[i + 7] > 0x7F) {
+    // process 4 bytes at a time when possible
+    for (; i + 3 < end; i += 4) {
+        const a = buffer[i], b = buffer[i + 1], c = buffer[i + 2], d = buffer[i + 3];
+
+        // Check all 4 bytes at once for non-ASCII
+        if ((a | b | c | d) & 0x80) {
             // Non-ASCII character detected, fall back to the generic utf8 implementation
             return utf8_decode_fallback(buffer, start, end, str, i);
         }
-        // Process 8 ASCII bytes at once
+        // Process 4 ASCII bytes at once
         str += String.fromCharCode(
-            buffer[i],
-            buffer[i + 1],
-            buffer[i + 2],
-            buffer[i + 3],
-            buffer[i + 4],
-            buffer[i + 5],
-            buffer[i + 6],
-            buffer[i + 7]
+            a, b, c, d
         );
     }