--- a/src/reader.js
+++ b/src/reader.js
@@ -327,6 +327,10 @@ Reader.prototype.bytes = function read_bytes() {
  * @returns {string} Value read
  */
 Reader.prototype.string = function read_string() {
+    // Note that we could simply use the `.bytes()` function. However, creating
+    // slices of a Uint8Array tends to be pretty expensive. If we instead just
+    // call utf8.read with appropriate start and end indicies, we can often
+    // avoid creating one of these slices and save some time.
     var length = this.uint32(),
         start  = this.pos,
         end    = this.pos + length;